% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaode-provider.R
\name{add_layer}
\alias{add_layer}
\title{Add Layer}
\usage{
add_layer(map, layer)
}
\arguments{
\item{map}{the map object}

\item{layer}{the layer to add}
}
\value{
the map object
}
\description{
Adds a layer to the map.
}
