% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baidu-provider.R
\name{baidu_js_dependency}
\alias{baidu_js_dependency}
\title{Create Baidu JavaScript Dependency}
\usage{
baidu_js_dependency()
}
\value{
htmltools dependency object
}
\description{
Create the JavaScript dependency for Baidu Maps integration.
}
\keyword{internal}
