% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative-sampler.R
\name{BoundaryValidator}
\alias{BoundaryValidator}
\title{Boundary Validator}
\description{
The BoundaryValidator class ensures that generated points
fall strictly within polygon boundaries.
}
\details{
Validates that points fall within polygon boundaries.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BoundaryValidator-validate_points}{\code{BoundaryValidator$validate_points()}}
\item \href{#method-BoundaryValidator-clone}{\code{BoundaryValidator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoundaryValidator-validate_points"></a>}}
\if{latex}{\out{\hypertarget{method-BoundaryValidator-validate_points}{}}}
\subsection{Method \code{validate_points()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoundaryValidator$validate_points(points, polygons)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{points}}{sf object with points}

\item{\code{polygons}}{sf object with polygons}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Validated points
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BoundaryValidator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BoundaryValidator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoundaryValidator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
