% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-switching.R
\name{check_authentication_compatibility}
\alias{check_authentication_compatibility}
\title{Check Authentication Compatibility}
\usage{
check_authentication_compatibility(from_capabilities, to_capabilities)
}
\arguments{
\item{from_capabilities}{List of source provider capabilities}

\item{to_capabilities}{List of target provider capabilities}
}
\value{
Character string describing issue or NULL if compatible
}
\description{
Check Authentication Compatibility
}
