% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-config.R
\name{create_default_provider_configs}
\alias{create_default_provider_configs}
\title{Create Default Provider Configurations}
\usage{
create_default_provider_configs()
}
\value{
Named list of ProviderConfig objects
}
\description{
Create default configurations for all supported providers.
}
\examples{
\donttest{
# Get default configurations
configs <- create_default_provider_configs()
mapbox_config <- configs$mapbox
}

}
