% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-switching.R
\name{extract_provider_info}
\alias{extract_provider_info}
\title{Extract Provider Information}
\usage{
extract_provider_info(map)
}
\arguments{
\item{map}{A mapdeck map object}
}
\value{
List containing provider information
}
\description{
Extract current provider information from a map object.
}
