% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-processor.R
\name{generate_points_in_polygon_worker}
\alias{generate_points_in_polygon_worker}
\title{Generate Points in Polygon Worker Function}
\usage{
generate_points_in_polygon_worker(polygon, n_samples, use_gpu)
}
\arguments{
\item{polygon}{sf object with a single polygon}

\item{n_samples}{Number of samples to generate}

\item{use_gpu}{Logical for GPU acceleration}
}
\value{
sf object with sampled points
}
\description{
Worker function for generating points within a polygon.
}
\keyword{internal}
