% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-transformer.R
\name{get_coordinate_detector}
\alias{get_coordinate_detector}
\title{Get Coordinate Detector}
\usage{
get_coordinate_detector()
}
\value{
CoordinateDetector instance
}
\description{
Get the global coordinate detector instance, creating it if necessary.
}
\examples{
\donttest{
# Get global detector
detector <- get_coordinate_detector()
result <- detector$detect(data)
}

}
