% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-switching.R
\name{get_coordinate_transformer}
\alias{get_coordinate_transformer}
\title{Get Coordinate Transformer}
\usage{
get_coordinate_transformer()
}
\value{
CoordinateTransformer instance
}
\description{
Get the global coordinate transformer instance, creating it if necessary.
}
\examples{
\donttest{
# Get global transformer
transformer <- get_coordinate_transformer()
coords <- transformer$transform(c(116.3974, 39.9093), "WGS84", "GCJ02")
}

}
