% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{get_default_style}
\alias{get_default_style}
\title{Get Default Style for Provider}
\usage{
get_default_style(provider)
}
\arguments{
\item{provider}{Character string identifying the provider}
}
\value{
Default style for the provider
}
\description{
Get Default Style for Provider
}
