% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-switching.R
\name{get_degraded_features}
\alias{get_degraded_features}
\title{Get Degraded Features}
\usage{
get_degraded_features(from_provider, to_provider)
}
\arguments{
\item{from_provider}{Character string identifying source provider}

\item{to_provider}{Character string identifying target provider}
}
\value{
Character vector of features that will be degraded
}
\description{
Get Degraded Features
}
