% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-factory.R
\name{get_provider_capabilities}
\alias{get_provider_capabilities}
\title{Get Provider Capabilities}
\usage{
get_provider_capabilities(provider_name)
}
\arguments{
\item{provider_name}{Character string identifying the provider}
}
\value{
List containing provider capabilities
}
\description{
Convenience function to get provider capabilities using the global factory.
}
\examples{
\donttest{
# Get Mapbox capabilities
capabilities <- get_provider_capabilities("mapbox")
}

}
