% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-switching.R
\name{get_provider_coordinate_system}
\alias{get_provider_coordinate_system}
\title{Get Provider Coordinate System}
\usage{
get_provider_coordinate_system(provider)
}
\arguments{
\item{provider}{Character string identifying the provider}
}
\value{
Character string identifying coordinate system
}
\description{
Get Provider Coordinate System
}
