% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_tokens.R
\name{get_provider_env_vars}
\alias{get_provider_env_vars}
\title{Get Provider Environment Variables}
\usage{
get_provider_env_vars(provider)
}
\arguments{
\item{provider}{Character string identifying the provider}
}
\value{
Character vector of environment variable names
}
\description{
Get list of environment variables to check for each provider
}
\examples{
\donttest{
# Get Mapbox environment variables
vars <- get_provider_env_vars("mapbox")
}

}
