% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-factory.R
\name{get_provider_factory}
\alias{get_provider_factory}
\title{Get Global Provider Factory}
\usage{
get_provider_factory()
}
\value{
ProviderFactory instance
}
\description{
Get the global provider factory instance, creating it if necessary.
}
\examples{
\donttest{
# Get global factory
factory <- get_provider_factory()
providers <- factory$get_available_providers()
}

}
