% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-management.R
\name{get_style_resolver}
\alias{get_style_resolver}
\title{Get Style Resolver}
\usage{
get_style_resolver()
}
\value{
StyleResolver instance
}
\description{
Get the global style resolver instance, creating it if necessary.
}
\examples{
\donttest{
# Get global style resolver
resolver <- get_style_resolver()
style <- resolver$resolve_style("streets", "mapbox")
}

}
