% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-management.R
\name{get_style_validator}
\alias{get_style_validator}
\title{Get Style Validator}
\usage{
get_style_validator()
}
\value{
StyleValidator instance
}
\description{
Get the global style validator instance, creating it if necessary.
}
\examples{
\donttest{
# Get global style validator
validator <- get_style_validator()
is_valid <- validator$validate_style("streets", "mapbox")
}

}
