% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-management.R
\name{get_theme_manager}
\alias{get_theme_manager}
\title{Get Theme Manager}
\usage{
get_theme_manager()
}
\value{
ThemeManager instance
}
\description{
Get the global theme manager instance, creating it if necessary.
}
\examples{
\donttest{
# Get global theme manager
theme_manager <- get_theme_manager()
theme_config <- theme_manager$apply_theme("dark", "mapbox")
}

}
