% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-store.R
\name{get_token_store}
\alias{get_token_store}
\title{Get Global Token Store}
\usage{
get_token_store(encrypted = FALSE)
}
\arguments{
\item{encrypted}{Logical indicating if tokens should be encrypted}
}
\value{
TokenStore instance
}
\description{
Get the global token store instance, creating it if necessary.
}
\examples{
\donttest{
# Get global token store
store <- get_token_store()
}

}
