% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-processor.R,
%   R/proportional-sampler.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y

x if not NULL, otherwise y
}
\description{
Null-coalescing operator

Null-coalescing operator
}
\keyword{internal}
