% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck_view}
\alias{mapdeck_view}
\title{Mapdeck view}
\usage{
mapdeck_view(
  map,
  location = NULL,
  zoom = NULL,
  pitch = NULL,
  bearing = NULL,
  duration = NULL,
  transition = c("linear", "fly")
)
}
\arguments{
\item{map}{a \code{mapdeck} map object}

\item{location}{Numeric vector of initial map center coordinates c(longitude, latitude).
Coordinates will be automatically transformed for Chinese providers if needed.}

\item{zoom}{Initial zoom level of the map (typically 0-20, varies by provider).}

\item{pitch}{The pitch angle of the map in degrees (0-60). Note: Not all providers
support pitch (e.g., Leaflet is limited to 0).}

\item{bearing}{Initial bearing/rotation of the map in degrees (0-360). Note: Not all
providers support bearing (e.g., Leaflet is limited to 0).}

\item{duration}{time in milliseconds of the transition}

\item{transition}{type of transition}
}
\description{
Changes the view of the of the map
}
