% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-switching.R
\name{merge_provider_configs}
\alias{merge_provider_configs}
\title{Merge Provider Configurations}
\usage{
merge_provider_configs(user_config, state_config)
}
\arguments{
\item{user_config}{List containing user-provided configuration}

\item{state_config}{List containing configuration from map state}
}
\value{
Merged configuration list
}
\description{
Merge Provider Configurations
}
