% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_baidu_style}
\alias{normalize_baidu_style}
\title{Normalize Baidu Style}
\usage{
normalize_baidu_style(style)
}
\arguments{
\item{style}{Character string containing Baidu style name}
}
\value{
Normalized Baidu style name
}
\description{
Normalize Baidu Style
}
