% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_character_style}
\alias{normalize_character_style}
\title{Normalize Character Style}
\usage{
normalize_character_style(style, provider)
}
\arguments{
\item{style}{Character string containing style name}

\item{provider}{Character string identifying the provider}
}
\value{
Normalized style string
}
\description{
Normalize Character Style
}
