% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_custom_style}
\alias{normalize_custom_style}
\title{Normalize Custom Style}
\usage{
normalize_custom_style(style, provider)
}
\arguments{
\item{style}{List containing custom style specification}

\item{provider}{Character string identifying the provider}
}
\value{
Normalized style list
}
\description{
Normalize Custom Style
}
