% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_gaode_style}
\alias{normalize_gaode_style}
\title{Normalize Gaode Style}
\usage{
normalize_gaode_style(style)
}
\arguments{
\item{style}{Character string containing Gaode style name}
}
\value{
Normalized Gaode style name
}
\description{
Normalize Gaode Style
}
