% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_leaflet_style}
\alias{normalize_leaflet_style}
\title{Normalize Leaflet Style}
\usage{
normalize_leaflet_style(style)
}
\arguments{
\item{style}{Character string containing Leaflet tile provider name}
}
\value{
Normalized Leaflet tile provider name
}
\description{
Normalize Leaflet Style
}
