% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_leaflet_tile_provider}
\alias{normalize_leaflet_tile_provider}
\title{Normalize Leaflet Tile Provider}
\usage{
normalize_leaflet_tile_provider(provider)
}
\arguments{
\item{provider}{Character string containing tile provider name}
}
\value{
Normalized tile provider name
}
\description{
Normalize Leaflet Tile Provider
}
