% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_mapbox_style}
\alias{normalize_mapbox_style}
\title{Normalize Mapbox Style}
\usage{
normalize_mapbox_style(style)
}
\arguments{
\item{style}{Character string containing Mapbox style name}
}
\value{
Normalized Mapbox style URL
}
\description{
Normalize Mapbox Style
}
