% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openlayers-provider.R
\name{normalize_openlayers_source}
\alias{normalize_openlayers_source}
\title{Normalize OpenLayers Source}
\usage{
normalize_openlayers_source(source)
}
\arguments{
\item{source}{Character string identifying the source}
}
\value{
Normalized source name
}
\description{
Normalize source names for OpenLayers.
}
\examples{
\donttest{
# Normalize source names
source <- normalize_openlayers_source("osm")
}

}
