% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_openlayers_style}
\alias{normalize_openlayers_style}
\title{Normalize OpenLayers Style}
\usage{
normalize_openlayers_style(style)
}
\arguments{
\item{style}{Character string containing OpenLayers source name}
}
\value{
Normalized OpenLayers source name
}
\description{
Normalize OpenLayers Style
}
