% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{normalize_style_name}
\alias{normalize_style_name}
\title{Normalize Style Name}
\usage{
normalize_style_name(style, provider)
}
\arguments{
\item{style}{Character string or list containing style specification}

\item{provider}{Character string identifying the provider}
}
\value{
Normalized style specification
}
\description{
Normalize style names for different providers.
}
\examples{
\donttest{
# Normalize style names
style <- normalize_style_name("dark", "mapbox")
}

}
