% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_baidu_layer}
\alias{validate_baidu_layer}
\title{Validate Baidu Layer}
\usage{
validate_baidu_layer(layer)
}
\arguments{
\item{layer}{List containing layer configuration}
}
\value{
Validated Baidu layer
}
\description{
Validate Baidu Layer
}
