% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_baidu_options}
\alias{validate_baidu_options}
\title{Validate Baidu Options}
\usage{
validate_baidu_options(options)
}
\arguments{
\item{options}{List containing options}
}
\value{
Validated Baidu options
}
\description{
Validate Baidu Options
}
