% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_gaode_layer}
\alias{validate_gaode_layer}
\title{Validate Gaode Layer}
\usage{
validate_gaode_layer(layer)
}
\arguments{
\item{layer}{List containing layer configuration}
}
\value{
Validated Gaode layer
}
\description{
Validate Gaode Layer
}
