% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_gaode_options}
\alias{validate_gaode_options}
\title{Validate Gaode Options}
\usage{
validate_gaode_options(options)
}
\arguments{
\item{options}{List containing options}
}
\value{
Validated Gaode options
}
\description{
Validate Gaode Options
}
