% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_leaflet_layer}
\alias{validate_leaflet_layer}
\title{Validate Leaflet Layer}
\usage{
validate_leaflet_layer(layer)
}
\arguments{
\item{layer}{List containing layer configuration}
}
\value{
Validated Leaflet layer
}
\description{
Validate Leaflet Layer
}
