% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_leaflet_options}
\alias{validate_leaflet_options}
\title{Validate Leaflet Options}
\usage{
validate_leaflet_options(options)
}
\arguments{
\item{options}{List containing options}
}
\value{
Validated Leaflet options
}
\description{
Validate Leaflet Options
}
