% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_mapbox_layer}
\alias{validate_mapbox_layer}
\title{Validate Mapbox Layer}
\usage{
validate_mapbox_layer(layer)
}
\arguments{
\item{layer}{List containing layer configuration}
}
\value{
Validated Mapbox layer
}
\description{
Validate Mapbox Layer
}
