% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_mapbox_options}
\alias{validate_mapbox_options}
\title{Validate Mapbox Options}
\usage{
validate_mapbox_options(options)
}
\arguments{
\item{options}{List containing options}
}
\value{
Validated Mapbox options
}
\description{
Validate Mapbox Options
}
