% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_openlayers_layer}
\alias{validate_openlayers_layer}
\title{Validate OpenLayers Layer}
\usage{
validate_openlayers_layer(layer)
}
\arguments{
\item{layer}{List containing layer configuration}
}
\value{
Validated OpenLayers layer
}
\description{
Validate OpenLayers Layer
}
