% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_openlayers_options}
\alias{validate_openlayers_options}
\title{Validate OpenLayers Options}
\usage{
validate_openlayers_options(options)
}
\arguments{
\item{options}{List containing options}
}
\value{
Validated OpenLayers options
}
\description{
Validate OpenLayers Options
}
