% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-interface.R
\name{validate_provider_interface}
\alias{validate_provider_interface}
\title{Validate Provider Interface Implementation}
\usage{
validate_provider_interface(provider_class)
}
\arguments{
\item{provider_class}{R6 class object to validate}
}
\value{
Logical indicating if implementation is valid
}
\description{
Utility function to validate that a provider class properly implements
the IMapProvider interface.
}
\examples{
\donttest{
# Validate a provider implementation
# is_valid <- validate_provider_interface(MapboxProvider)
}

}
