% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-utils.R
\name{validate_provider_specific_layer}
\alias{validate_provider_specific_layer}
\title{Validate Provider-Specific Layer}
\usage{
validate_provider_specific_layer(layer, provider)
}
\arguments{
\item{layer}{List containing layer configuration}

\item{provider}{Character string identifying the provider}
}
\value{
Validated layer configuration
}
\description{
Validate Provider-Specific Layer
}
